Added all the Primary IDs to Autoincrement and userId changed to userEmail to make login userEmail based

+++ CRS-GROUPD-WIBMO-POS/src/com/wibmo/dao/CourseDAO.java
@@ -8,5 +8,5 @@ package com.wibmo.dao;
  */
 public interface CourseDAO {
 
-	public boolean searchCourse(long courseId);
+	public boolean searchCourse(int courseId);
 }
